airflow_machine/requirements.txt
# Airflow Machine Requirements

# Core
apache-airflow==2.7.0
pyspark>=3.5.0

# Kafka
kafka-python>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
requests>=2.31.0

# Database (choose one)
# PostgreSQL
psycopg2-binary>=2.9.0
# MySQL
# mysql-connector-python>=8.0.0

